Remove linenote-browse.
Using a transient, it's easy to jump from note to note, and Eldoc should
display the note's contents, so I don't think this functionality is really
needed.

I may decide to bring it back later, though, if it turns out to be useful,
e.g., for perusing all notes, not just the ones in the current buffer.