Refactor `linenote--create-linenum-string`.
Do not take an optional argument. Instead, check if there's a linenote
overlay at point and use the line numbers stored in it. Otherwise, use the
region if it's active or point.

This way, if there's already a note for the relevant line, we can retrieve
it, and if there's not, we can create a new one.