Add `line` argument to `linenote--get-note-at-point`.
In order to get `linenote-{previous+next}-note` right, we need to
distinguish between getting the note at the current line and checking if
there is a note at point. The problem is that these functions use
`{previous|next}-single-char-property-change` to get out of the current
note, but when point is at the end of a line, it's no longer inside a note
and there's no need to move out of it.